requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
python-dotenv==1.0.0
pymongo[srv]==4.6.1
redis==5.0.1
supabase==1.2.0
python-jose[cryptography]==3.4.1
google-cloud-secret-manager==2.18.0
sentence-transformers==2.2.2
pandas==2.1.4
scikit-learn==1.3.2
google-cloud-storage==2.14.0